requirements_txt.txt
streamlit==1.32.0
google-generativeai==0.4.0
pymupdf==1.23.26
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
torch==2.2.1
transformers==4.38.2
